LPS-55928 add specific interface for reindexing ddm
